Implement login features on boot app

+++ backend-boot/src/main/java/com/example/demo/controller/TodoController.java
@@ -2,10 +2,16 @@ package com.example.demo.controller;
 
 import com.example.demo.entity.Todo;
 import com.example.demo.service.TodoService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 @RestController
 public class TodoController {
+
+    @Autowired
+    private UserSession userSession;
 
     private final TodoService todoService;
 
@@ -17,6 +23,9 @@ public class TodoController {
     @CrossOrigin(origins = {"http://localhost:300"}, methods = RequestMethod.PUT)
 //    @CrossOrigin(origins = {"http://localhost:3000"})
     public Todo put(@PathVariable Long todoId) {
+        if (userSession.getUserId() == null) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "forbidden");
+        }
         return todoService.updateStatus(todoId);
     }
 
@@ -24,6 +33,9 @@ public class TodoController {
     @CrossOrigin
     //@CrossOrigin(origins = {"http://localhost:3000"})
     public void delete(@PathVariable Long todoId) {
+        if (userSession.getUserId() == null) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "forbidden");
+        }
         todoService.removeTodo(todoId);
     }
 }
